fix: htons shenanigans.
- htons and ntohs get aliased to a macro on some systems.
This creates all sorts of namespace collision nightmares, but
doesn't actually resolve the problem of needing htons and ntohs
to explicitly be constexpr. Instead, I have renamed the internal
implementations of htons and ntohs to htons_ and ntohs_ and don't
use them anywhere I don't explicitly need a constexpr declaration.